JE-105: Add getAllInvoiceEntries API call and adjust numerical types

+++ bundles/Billing/Entity/InvoiceEntry.php
@@ -60,7 +60,7 @@ class InvoiceEntry
     private $product;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="decimal", precision=10, scale=2)
      */
     private $price;
 
@@ -170,12 +170,12 @@ class InvoiceEntry
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
